% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{calculate_log_likelihood}
\alias{calculate_log_likelihood}
\title{Calculate log likelihood of data set}
\usage{
calculate_log_likelihood(
  constraint_weights,
  input_file,
  bias_file = NA,
  mu_scalar = NA,
  mu_vector = NA,
  sigma_scalar = NA,
  sigma_vector = NA,
  input_format = "otsoft",
  in_sep = "\\t"
)
}
\arguments{
\item{input_file}{The path to the input data file. This file contains one or
more OT tableaux consisting of mappings between underlying and surface
forms with observed frequency and violation profiles. Constraint
violations must be numeric.}

\item{bias_file}{(optional) The path to the file containing mus and sigma
for constraint biases. If this argument is provided, the scalar and vector
mu and sigma arguments will be ignored.}

\item{mu_scalar}{(optional) A single scalar value that will serve as the mu
for each constraint in the bias term. Constraint weights will also be
initialized to this value. This value will not be used if either
\code{bias_file} or \code{mu_vector} are provided.}

\item{mu_vector}{(optional) A vector of mus for each constraint in the bias
term. The length of this vector must equal the number of constraints in
the input file. If \code{bias_file} is provided, this argument will be
ignored. If this argument is provided, \code{mu_scalar} will be ignored.}

\item{sigma_scalar}{(optional) A single scalar value that will serve as the
sigma for each constraint in the bias term. This value will not be used if
either \code{bias_file} or \code{sigma_vector} are provided.}

\item{sigma_vector}{(optional) A vector of sigmas for each constraint in the
bias term. The length of this vector must equal the number of constraints
in the input file. If \code{bias_file} is provided, this argument will be ignored.
If this argument is provided, \code{sigma_scalar} will be ignored.}

\item{input_format}{(optional) A string specifying the format of the input
files. Currently only OTSoft-style formatting is supported.}

\item{in_sep}{(optional) The delimiter used in the input files. Defaults to
tabs.}
}
\value{
The log likelihood of the data set.
}
\description{
Calculates the log likelihood of a set of tableaux given a set of constraint
weights and optional biases. This is calculated using the same equations
described in the documentation for \code{optimize_weights}, but it simply uses
the provided constraint weights rather than finding optimal ones.
}
\details{
If no bias parameters are specified (either the \code{bias_file} argument or some
combination of the scalar/vector mu/sigma parameters), optimization will be
done without the bias term.
}
