% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_constraints.R
\name{compare_aic}
\alias{compare_aic}
\title{Compares given model using aic, aic_c, or bic (based on user choice)
with compare_models}
\usage{
compare_aic(model, candidate_models, candidate_constraints, method)
}
\description{
Compares given model using aic, aic_c, or bic (based on user choice)
with compare_models
}
