% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{optimize_fold}
\alias{optimize_fold}
\title{Helper function that takes a bias vector, potential mu/sigma vector
(if mapping to each value), train/test splits,
outputs the test loglik to be used in optim later}
\usage{
optimize_fold(par, train_data, test_data)
}
\description{
Helper function that takes a bias vector, potential mu/sigma vector
(if mapping to each value), train/test splits,
outputs the test loglik to be used in optim later
}
