% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize_weights}
\alias{optimize_weights}
\title{Optimizes constraint weights given a data set and optional biases. If no
bias arguments are provided, the bias term will not be included in the
optimization.}
\usage{
optimize_weights(
  input_file,
  bias_file = NA,
  mu_scalar = NA,
  mu_vector = NA,
  sigma_scalar = NA,
  sigma_vector = NA,
  input_format = "otsoft"
)
}
\arguments{
\item{input_file}{The path to the input data file. This file contains one or
more OT tableaux consisting of mappings between underlying and surface
forms with observed frequency and violation profiles. Constraint
violations must be numeric.}

\item{bias_file}{(optional) The path to the file containing mus and sigma
for constraint biases. If this argument is provided, the scalar and vector
mu and sigma arguments will be ignored.}

\item{mu_scalar}{(optional) A single scalar value that will serve as the mu
for each constraint in the bias term. Constraint weights will also be
initialized to this value. This value will not be used if either bias_file
or mu_vector are provided.}

\item{mu_vector}{(optional) A vector of mus for each constraint in the bias
term. The length of this vector must equal the number of constraints in
the input file. If bias_file is provided, this argument will be ignored.
If this argument is provided, mu_scalar will be ignored.}

\item{sigma_scalar}{(optional) A single scalar value that will serve as the
sigma for each constraint in the bias term. This value will not be used if
either bias_file or sigma_vector are provided.}

\item{sigma_vector}{(optional) A vector of sigmas for each constraint in the
bias term. The length of this vector must equal the number of constraints
in the input file. If bias_file is provided, this argument will be ignored.
If this argument is provided, sigma_scalar will be ignored.}

\item{input_format}{(optional) A string specifying the format of the input
files. Currently only otsoft-style formatting is supported.}
}
\value{
A named vector containing optimized constraint weights.
}
\description{
TODO: Add description of how optimization functions.
}
\examples{
  optimize_weights('my_tableaux.csv')
  optimize_weights('my_tableaux.csv', 'my_biases.csv')
  optimize_weights('my_tableaux.csv', mu_vector = c(1,2), sigma_vector = c(100, 200))
  optimize_weights('my_tableaux.csv', mu_scalar = 0, sigma_scalar = 1000)
  optimize_weights('my_tableaux.csv', mu_vector = c(1, 2), sigma_scalar = 1000)

}
