% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_constraints.R
\name{step}
\alias{step}
\title{Recursively generates all subsets of weights and fits optimal models to each
subset, ultimately returning all models (to be used in minimal_constraints)}
\usage{
step(dataset, model, weights)
}
\description{
Recursively generates all subsets of weights and fits optimal models to each
subset, ultimately returning all models (to be used in minimal_constraints)
}
