% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_constraints.R
\name{step_naive}
\alias{step_naive}
\title{Iteratively drops one constraint at a time (naive approach)}
\usage{
step_naive(dataset, method = "aic")
}
\description{
Iteratively drops one constraint at a time (naive approach)
}
