% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_constraints.R
\name{minimal_constraints}
\alias{minimal_constraints}
\title{Find the optimal minimal subset of weights for a given dataset and model choice.}
\usage{
minimal_constraints(dataset, model, weights)
}
\description{
Uses stepwise regression with backwards elimination to selectively remove
suboptimal weights as to maximize model performance (through log-likelihood)
while minimizing the number of weights.
}
