% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_constraints.R
\name{minimal_constraints}
\alias{minimal_constraints}
\title{Find the optimal minimal subset of weights for a given dataset and model choice.}
\usage{
minimal_constraints(
  dataset,
  method = "aic",
  approach = "iterative",
  max_depth = Inf
)
}
\description{
Uses stepwise regression with backwards elimination to find all subsets of
weights and compares these subset models to one another.
}
