% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_probabilities}
\alias{predict_probabilities}
\title{Predict probabilities of candidates based on their violation profiles and
constraint weights.}
\usage{
predict_probabilities(test_file, constraint_weights, input_format = "otsoft")
}
\arguments{
\item{test_file}{The path to the input data file. This file contains one or
more OT tableaux consisting of mappings between underlying and surface
forms with observed frequency and violation profiles. Constraint
violations must be numeric.}

\item{constraint_weights}{The constraint weights to use.}

\item{input_format}{(optional) A string specifying the format of the input
files. Currently only otsoft-style formatting is supported. Note that this
means a column containing frequencies must be present. However, these
frequencies will be ignored.}
}
\value{
A data table containing all the tableaux, with probabilities
assigned to each candidate.
}
\description{
Predict probabilities of candidates based on their violation profiles and
constraint weights.
}
