% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_probabilities}
\alias{predict_probabilities}
\title{Predict probabilities of OT candidates}
\usage{
predict_probabilities(
  test_file,
  constraint_weights,
  output_path = NA,
  input_format = "otsoft",
  in_sep = "\\t",
  out_sep = "\\t"
)
}
\arguments{
\item{test_file}{The path to the input data file. This file contains one or
more OT tableaux consisting of mappings between underlying and surface
forms with observed frequency and violation profiles. Constraint
violations must be numeric.}

\item{constraint_weights}{The constraint weights to use.}

\item{output_path}{(optional) A string specifying the path to a file to
which the output will be saved. If the file exists it will be overwritten.
If this argument isn't provided, the output will not be written to a file.}

\item{input_format}{(optional) A string specifying the format of the input
files. Currently only otsoft-style formatting is supported. Note that this
means a column containing frequencies must be present. However, these
frequencies will be ignored.}

\item{in_sep}{(optional) The delimiter used in the input file.
Defaults to tabs.}

\item{out_sep}{(optional) The delimiter used in the output files.
Defaults to tabs.}
}
\value{
A data table containing all the tableaux, with probabilities
assigned to each candidate.
}
\description{
Predict probabilities of candidates based on their violation profiles and
constraint weights.
}
\details{
For each input/output pair in the provided, file this
will calculate the probability of that output given the input form and the
provided weights. This is defined as

\deqn{P(y|x; w) = \frac{1}{Z_w(x)}\exp(\sum_{k=1}^{m}{w_k f_k(y, x)})}

where \eqn{f_k(y, x)} is the number of violatons of constraint \eqn{k}
incurred by mapping underlying \eqn{x} to surface \eqn{y}, \eqn{w_k} is the
weight associated with contraint \eqn{k}, and  \eqn{Z(x)} is a normalization
term defined as

\deqn{Z(x) = \sum_{y\in \mathcal{Y}(x)}{\exp(\sum_{k=1}^{m}{w_k f_k(y, x)})}}

The resulting probabilities will be appended to a data frame object
representing the input tableaux. This data frame can also be saved to a file
if the \code{output_path} argument is provided.
}
